remove TransferHelper from LimitSwapVerifier
the checks that TransferHelper does aren't needed. The final requirement that the limit swap output amount is transferred to the account ensures that the swap worked as expected. It's up to the code executed via CALL_EXECUTOR.proxyCall() to check that input asset transfer worked properly before proving the output asset.

it's also more gas optimized to remove the TransferHelper wrapper functions